rename 'metrics' folder to 'results'
