--- docs/source/api_reference/metrics/disease.rst
@@ -1,1 +1,0 @@
-.. automodule:: vivarium_public_health.metrics.disease
