+++ docs/source/api_reference/results/risk.rst
@@ -1,0 +1,1 @@
+.. automodule:: vivarium_public_health.results.risk
