+++ docs/source/api_reference/results/stratification.rst
@@ -1,0 +1,1 @@
+.. automodule:: vivarium_public_health.results.stratification
