+++ src/vivarium_public_health/results/risk.py
@@ -13,7 +13,7 @@ import pandas as pd
 from vivarium.framework.engine import Builder
 from vivarium.framework.results import StratifiedObserver
 
-from vivarium_public_health.metrics.columns import COLUMNS
+from vivarium_public_health.results.columns import COLUMNS
 from vivarium_public_health.utilities import to_years
