+++ src/vivarium_public_health/results/stratification.py
