+++ tests/results/test_disease_observer.py
@@ -9,9 +9,9 @@ from vivarium.testing_utilities import TestPopulation, build_table
 
 from vivarium_public_health.disease import DiseaseModel, DiseaseState
 from vivarium_public_health.disease.state import SusceptibleState
-from vivarium_public_health.metrics.columns import COLUMNS
-from vivarium_public_health.metrics.disease import DiseaseObserver
-from vivarium_public_health.metrics.stratification import ResultsStratifier
+from vivarium_public_health.results.columns import COLUMNS
+from vivarium_public_health.results.disease import DiseaseObserver
+from vivarium_public_health.results.stratification import ResultsStratifier
 from vivarium_public_health.utilities import to_years
